Banking App 2 fully functional, still need to implement JUnit testing and documentation

+++ BankingAppIteration2/src/main/java/com/revature/dao/TransactionDao.java
@@ -6,8 +6,8 @@ import com.revature.pojos.Transaction;
 
 public interface TransactionDao {
 	public List<Transaction> getTransactions ();
-	public List<Transaction> getAccountTransactions(int acctNumber);
-	public List<Transaction> getAccountTransactions(int acctNumber, int max);
+	public List<Transaction> getAccountTransactions(long acctNumber);
+	public List<Transaction> getAccountTransactions(long acctNumber, int max);
 	public int createTransaction(Transaction action);
 	
 	//won't need for this version of the app but may be useful in the future
